moving patron code up the component hierarchy
